task_tracker_project/requirements.txt
Django>=5.2
djangorestframework
psycopg2-binary
djangorestframework-simplejwt

